// src/types/collection.ts
export type CollectionFilterAppliedInterface = {
  __typename?: string;
  code: string;
  label: string;
  includeInFilters: boolean;
  renderer?: string | null;
  type: string;
};

export type CollectionFilterAvailableInterface = {
  __typename?: string;
  code: string;
  label: string;
  includeInFilters: boolean;
  renderer?: string | null;
  type: string;
};
// ...
export type CollectionFilterAppliedBoolean =
  CollectionFilterAppliedInterface & {
    __typename?: 'CollectionFilterAppliedBoolean';
    appliedValue: boolean;
    trueLabel: string;
    falseLabel: string;
  };

export type CollectionFilterAvailableBoolean =
  CollectionFilterAvailableInterface & {
    __typename?: 'CollectionFilterAvailableBoolean';
    trueItemCount?: number | null;
    falseItemCount?: number | null;
    trueLabel: string;
    falseLabel: string;
  };